requirements.txt
# Base
pytz==2018.7
python-slugify==1.2.6

# Django
django==3.2.5

# Environment
django-environ==0.4.5

# Django Rest Framework
djangorestframework==3.13.1

# Static files
whitenoise==4.1.2

# JWT
pyjwt==2.4.0

# Passwords security
argon2-cffi==20.1.0

# Tools
django-extensions==3.2.0
